@@ -6,18 +6,24 @@
   "repository": "git://github.com/GitbookIO/slate-edit-list.git",
   "main": "./dist/index.js",
   "peerDependencies": {
-    "slate": "^0.30.0"
+    "immutable": "^3.8.2",
+    "slate": "^0.29.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.11.4",
-    "babel-core": "^6.11.4",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
-    "babelify": "^7.3.0",
-    "browserify": "^13.1.0",
-    "eslint": "^3.1.1",
-    "eslint-config-gitbook": "1.1.2",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.2",
+    "babel-plugin-transform-flow-strip-types": "^6.22.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babelify": "^8.0.0",
+    "browserify": "^13.3.0",
+    "eslint": "^4.10.0",
+    "eslint-config-gitbook": "^2.0.3",
+    "eslint-plugin-import": "^2.8.0",
     "expect": "^1.20.2",
+    "flow-bin": "^0.57.3",
     "gh-pages": "^0.11.0",
     "http-server": "^0.9.0",
     "immutable": "^3.8.2",
@@ -25,7 +31,7 @@
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "read-metadata": "^1.0.0",
-    "slate": "^0.30.0",
+    "slate": "^0.29.0",
     "slate-hyperscript": "^0.4.6",
     "slate-react": "^0.10.0",
     "stringify": "^5.1.0"
@@ -34,7 +40,7 @@
     "prepublish": "babel ./lib --out-dir ./dist",
     "postpublish": "npm run deploy-example",
     "lint": "eslint ./",
-    "build-example": "browserify ./example/main.js -o ./example/bundle.js -t [ babelify --presets [ es2015 react ] ] -t [ stringify --extensions [.yaml] ]",
+    "build-example": "browserify ./example/main.js -o ./example/bundle.js -t [ babelify --presets [ es2015 react stage-0 ] ] -t [ stringify --extensions [.yaml] ]",
     "serve-example": "http-server ./example/ -p 8081",
     "start": "npm run build-example && npm run serve-example",
     "deploy-example": "npm run build-example && gh-pages -d ./example",
