@@ -1,43 +1,47 @@
-const Slate = require('slate');
+// @flow
+import { Block, type Change } from 'slate';
 
-const getPreviousItem = require('../getPreviousItem');
-const getCurrentItem = require('../getCurrentItem');
-const getListForItem = require('../getListForItem');
-const isList = require('../isList');
+import type Options from '../options';
+import {
+    getPreviousItem,
+    getCurrentItem,
+    getListForItem,
+    isList
+} from '../utils';
 
 /**
  * Increase the depth of the current item by putting it in a sub-list
  * of previous item.
  * For first items in a list, does nothing.
- *
- * @param  {PluginOptions} opts
- * @param  {Slate.Change} change
- * @return {Slate.Change}
  */
-function increaseItemDepth(opts, change) {
+function increaseItemDepth(opts: Options, change: Change): Change {
     const previousItem = getPreviousItem(opts, change.value);
+    const currentItem = getCurrentItem(opts, change.value);
 
     if (!previousItem) {
         return change;
     }
 
-    const currentItem = getCurrentItem(opts, change.value);
+    if (!currentItem) {
+        return change;
+    }
 
     // Move the item in the sublist of previous item
     return moveAsSubItem(opts, change, currentItem, previousItem.key);
 }
 
 /**
- * Move the given item to the list at the end of destination node,
- * creating one if needed.
- *
- * @param {PluginOptions} opts
- * @param {Slate.Change} change
- * @param {Slate.Block} item The list item to add
- * @param {String} destKey The key of the destination node
- * @return {Slate.Change}
+ * Move the given item to the sublist at the end of destination item,
+ * creating a sublist if needed.
  */
-function moveAsSubItem(opts, change, item, destKey) {
+function moveAsSubItem(
+    opts: Options,
+    change: Change,
+    // The list item to add
+    item: Block,
+    // The key of the destination node
+    destKey: string
+): Change {
     const destination = change.value.document.getDescendant(destKey);
     const lastIndex = destination.nodes.size;
     const lastChild = destination.nodes.last();
@@ -51,28 +55,23 @@ function moveAsSubItem(opts, change, item, destKey) {
             existingList.key,
             existingList.nodes.size // as last item
         );
-    } else {
-        const currentList = getListForItem(opts, change.value, destination);
-
-        const newSublist = Slate.Block.create({
-            kind: 'block',
-            type: currentList.type,
-            data: currentList.data
-        });
-
-        change.insertNodeByKey(
-            destKey,
-            lastIndex,
-            newSublist,
-            { normalize: false }
-        );
-
-        return change.moveNodeByKey(
-            item.key,
-            newSublist.key,
-            0
-        );
     }
+    const currentList = getListForItem(opts, change.value, destination);
+    if (!currentList) {
+        throw new Error('Destination is not in a list');
+    }
+
+    const newSublist = Block.create({
+        kind: 'block',
+        type: currentList.type,
+        data: currentList.data
+    });
+
+    change.insertNodeByKey(destKey, lastIndex, newSublist, {
+        normalize: false
+    });
+
+    return change.moveNodeByKey(item.key, newSublist.key, 0);
 }
 
-module.exports = increaseItemDepth;
+export default increaseItemDepth;
