@@ -1,13 +1,13 @@
-const getItemsAtRange = require('../getItemsAtRange');
+// @flow
+import { type Change } from 'slate';
+
+import type Options from '../options';
+import { getItemsAtRange } from '../utils';
 
 /**
  * Unwrap items at range from their list.
- *
- * @param  {PluginOptions} opts
- * @param  {Slate.Change} change
- * @return {Change}
  */
-function unwrapList(opts, change) {
+function unwrapList(opts: Options, change: Change): Change {
     const items = getItemsAtRange(opts, change.value);
 
     if (items.isEmpty()) {
@@ -15,7 +15,9 @@ function unwrapList(opts, change) {
     }
 
     // Unwrap the items from their list
-    items.forEach(item => change.unwrapNodeByKey(item.key, { normalize: false }));
+    items.forEach(item =>
+        change.unwrapNodeByKey(item.key, { normalize: false })
+    );
 
     // Parent of the list of the items
     const firstItem = items.first();
@@ -26,15 +28,19 @@ function unwrapList(opts, change) {
     // Unwrap the items' children
     items.forEach(item => {
         item.nodes.forEach(node => {
-            change.moveNodeByKey(node.key, parent.key, index, { normalize: false });
-            index++;
+            change.moveNodeByKey(node.key, parent.key, index, {
+                normalize: false
+            });
+            index += 1;
         });
     });
 
     // Finally, remove the now empty items
-    items.forEach(item => change.removeNodeByKey(item.key));
+    items.forEach(item =>
+        change.removeNodeByKey(item.key, { normalize: false })
+    );
 
     return change;
 }
 
-module.exports = unwrapList;
+export default unwrapList;
