@@ -1,34 +1,43 @@
-const Slate = require('slate');
-const { List } = require('immutable');
-const isList = require('../isList');
+// @flow
+import { Data, type Value, type Change, type Block } from 'slate';
+import { List } from 'immutable';
+
+import type Options from '../options';
+import { isList } from '../utils';
 
 /**
  * Wrap the blocks in the current selection in a new list. Selected
  * lists are merged together.
- *
- * @param  {PluginOptions} opts
- * @param  {Slate.Change}
- * @param  {String?} type
- * @param  {Object|Data?} [data]
- * @return {Change}
  */
-function wrapInList(opts, change, ordered, data) {
+function wrapInList(
+    opts: Options,
+    change: Change,
+    type?: string,
+    data?: Object | Data
+): Change {
     const selectedBlocks = getHighestSelectedBlocks(change.value);
-    const type = ordered || opts.types[0];
+    type = type || opts.types[0];
 
     // Wrap in container
-    change.wrapBlock({
-        type,
-        data: Slate.Data.create(data)
-    });
+    change.wrapBlock(
+        {
+            type,
+            data: Data.create(data)
+        },
+        { normalize: false }
+    );
 
     // Wrap in list items
-    selectedBlocks.forEach((node) => {
+    selectedBlocks.forEach(node => {
         if (isList(opts, node)) {
             // Merge its items with the created list
-            node.nodes.forEach(({ key }) => change.unwrapNodeByKey(key));
+            node.nodes.forEach(({ key }) =>
+                change.unwrapNodeByKey(key, { normalize: false })
+            );
         } else {
-            change.wrapBlockByKey(node.key, opts.typeItem);
+            change.wrapBlockByKey(node.key, opts.typeItem, {
+                normalize: false
+            });
         }
     });
 
@@ -36,11 +45,9 @@ function wrapInList(opts, change, ordered, data) {
 }
 
 /**
- * @param  {Slate.Value} value
- * @return {List<Block>} The highest list of blocks that cover the
- * current selection
+ * Returns the highest list of blocks that cover the current selection
  */
-function getHighestSelectedBlocks(value) {
+function getHighestSelectedBlocks(value: Value): List<Block> {
     const range = value.selection;
     const { document } = value;
 
@@ -49,13 +56,12 @@ function getHighestSelectedBlocks(value) {
 
     if (startBlock === endBlock) {
         return List([startBlock]);
-    } else {
-        const ancestor = document.getCommonAncestor(startBlock.key, endBlock.key);
-        const startPath = ancestor.getPath(startBlock.key);
-        const endPath = ancestor.getPath(endBlock.key);
-
-        return ancestor.nodes.slice(startPath[0], endPath[0] + 1);
     }
+    const ancestor = document.getCommonAncestor(startBlock.key, endBlock.key);
+    const startPath = ancestor.getPath(startBlock.key);
+    const endPath = ancestor.getPath(endBlock.key);
+
+    return ancestor.nodes.slice(startPath[0], endPath[0] + 1);
 }
 
-module.exports = wrapInList;
+export default wrapInList;
