@@ -1,8 +1,9 @@
-const unwrapList = require('./changes/unwrapList');
-const splitListItem = require('./changes/splitListItem');
-const decreaseItemDepth = require('./changes/decreaseItemDepth');
-const getCurrentItem = require('./getCurrentItem');
-const getItemDepth = require('./getItemDepth');
+// @flow
+import { type Change } from 'slate';
+
+import type Options from '../options';
+import { unwrapList, splitListItem, decreaseItemDepth } from '../changes';
+import { getCurrentItem, getItemDepth } from '../utils';
 
 /**
  * User pressed Enter in an editor
@@ -11,11 +12,16 @@ const getItemDepth = require('./getItemDepth');
  * Enter in an empty list item should remove it
  * Shift+Enter in a list item should make a new line
  */
-function onEnter(event, change, editor, opts) {
+function onEnter(
+    event: *,
+    change: Change,
+    editor: *,
+    opts: Options
+): void | any {
     // Pressing Shift+Enter
     // should split block normally
     if (event.shiftKey) {
-        return null;
+        return undefined;
     }
 
     const { value } = change;
@@ -23,7 +29,7 @@ function onEnter(event, change, editor, opts) {
 
     // Not in a list
     if (!currentItem) {
-        return null;
+        return undefined;
     }
 
     event.preventDefault();
@@ -31,15 +37,12 @@ function onEnter(event, change, editor, opts) {
         // Block is empty, we exit the list
         if (getItemDepth(opts, value) > 1) {
             return decreaseItemDepth(opts, change);
-        } else {
-            // Exit list
-            return unwrapList(opts, change);
         }
-    } else {
-        // Split list item
-        return splitListItem(opts, change);
+        // Exit list
+        return unwrapList(opts, change);
     }
-
+    // Split list item
+    return splitListItem(opts, change);
 }
 
-module.exports = onEnter;
+export default onEnter;
