@@ -1,37 +1,34 @@
-const decreaseItemDepth = require('./changes/decreaseItemDepth');
-const increaseItemDepth = require('./changes/increaseItemDepth');
-const getCurrentItem = require('./getCurrentItem');
+// @flow
+import { type Change } from 'slate';
+
+import type Options from '../options';
+import { decreaseItemDepth, increaseItemDepth } from '../changes';
+import { getCurrentItem } from '../utils';
 
 /**
  * User pressed Tab in an editor.
  * Tab       -> Increase item depth if inside a list item
  * Shift+Tab -> Decrease item depth if inside a list item
  */
-function onTab(event, change, editor, opts) {
+function onTab(event: *, change: Change, editor: *, opts: Options): void | any {
     const { value } = change;
     const { isCollapsed } = value;
 
     if (!isCollapsed || !getCurrentItem(opts, value)) {
-        return;
+        return undefined;
     }
 
     // Shift+tab reduce depth
     if (event.shiftKey) {
         event.preventDefault();
 
-        return decreaseItemDepth(
-            opts,
-            change
-        );
+        return decreaseItemDepth(opts, change);
     }
 
     // Tab increases depth
     event.preventDefault();
 
-    return increaseItemDepth(
-        opts,
-        change
-    );
+    return increaseItemDepth(opts, change);
 }
 
-module.exports = onTab;
+export default onTab;
