@@ -1,17 +1,26 @@
-const { Record } = require('immutable');
+// @flow
+import { Record } from 'immutable';
 
-const DEFAULTS = {
-    // The possibles types for list containers
-    types: ['ul_list', 'ol_list'],
-    // The type of list items
-    typeItem: 'list_item',
-    // The type of default block in items
-    typeDefault: 'paragraph'
+export type OptionsFormat = {
+    types?: string[],
+    typeItem?: string,
+    typeDefault?: string
 };
 
 /**
  * The plugin options
  */
-class Options extends new Record(DEFAULTS) {}
+class Options extends Record({
+    types: ['ul_list', 'ol_list'],
+    typeItem: 'list_item',
+    typeDefault: 'paragraph'
+}) {
+    // The possibles types for list containers
+    types: string[];
+    // The type of list items
+    typeItem: string;
+    // The type of default block in items
+    typeDefault: string;
+}
 
-module.exports = Options;
+export default Options;
