@@ -1,14 +1,13 @@
-const getCurrentItem = require('./getCurrentItem');
+// @flow
+import { type Value, type Block } from 'slate';
+
+import type Options from '../options';
+import getCurrentItem from './getCurrentItem';
 
 /**
  * Get depth of current block in a document list
- *
- * @param {PluginOptions} opts
- * @param {Slate.Value} value
- * @param {Slate.Block} block?
- * @return {Number}
  */
-function getItemDepth(opts, value, block) {
+function getItemDepth(opts: Options, value: Value, block?: Block): number {
     const { document, startBlock } = value;
     block = block || startBlock;
 
@@ -19,7 +18,7 @@ function getItemDepth(opts, value, block) {
 
     const list = document.getParent(currentItem.key);
 
-    return (1 + getItemDepth(opts, value, list));
+    return 1 + getItemDepth(opts, value, list);
 }
 
-module.exports = getItemDepth;
+export default getItemDepth;
