@@ -1,17 +1,22 @@
-const { List } = require('immutable');
-const isList = require('./isList');
-const getCurrentItem = require('./getCurrentItem');
+// @flow
+import { type Value, type Block, type Range } from 'slate';
+import { List } from 'immutable';
+
+import type Options from '../options';
+import isList from './isList';
+import getCurrentItem from './getCurrentItem';
 
 /**
  * Return the list of items at the given range. The returned items are
  * the highest list item blocks that cover the range.
  *
- * @param {PluginOptions} opts
- * @param {Slate.Node} value
- * @param {Slate.Selection} [range]
- * @return {List<Slate.Block>} Empty if no list of items can cover the range
+ * Returns an empty list if no list of items can cover the range
  */
-function getItemsAtRange(opts, value, range) {
+function getItemsAtRange(
+    opts: Options,
+    value: Value,
+    range?: Range
+): List<Block> {
     range = range || value.selection;
 
     if (!range.startKey) {
@@ -25,9 +30,7 @@ function getItemsAtRange(opts, value, range) {
 
     if (startBlock === endBlock) {
         const item = getCurrentItem(opts, value, startBlock);
-        return item
-            ? List([item])
-            : List();
+        return item ? List([item]) : List();
     }
 
     const ancestor = document.getCommonAncestor(startBlock.key, endBlock.key);
@@ -40,10 +43,9 @@ function getItemsAtRange(opts, value, range) {
     } else if (ancestor.type === opts.typeItem) {
         // The ancestor is the highest list item that covers the range
         return List([ancestor]);
-    } else {
-        // No list of items can cover the range
-        return List();
     }
+    // No list of items can cover the range
+    return List();
 }
 
-module.exports = getItemsAtRange;
+export default getItemsAtRange;
