@@ -1,18 +1,20 @@
-const getCurrentItem = require('./getCurrentItem');
+// @flow
+import { type Value, type Block } from 'slate';
+
+import type Options from '../options';
+import getCurrentItem from './getCurrentItem';
 
 /**
  * Return the previous item, from current selection or from a node.
- *
- * @param {PluginOptions} opts
- * @param {Slate.Value} value
- * @param {Slate.Block} block?
- * @return {Slate.Block || Void}
  */
-function getPreviousItem(opts, value, block) {
+function getPreviousItem(opts: Options, value: Value, block?: Block): ?Block {
     const { document, startBlock } = value;
     block = block || startBlock;
 
     const currentItem = getCurrentItem(opts, value, block);
+    if (!currentItem) {
+        return null;
+    }
 
     const previousSibling = document.getPreviousSibling(currentItem.key);
 
@@ -20,9 +22,8 @@ function getPreviousItem(opts, value, block) {
         return null;
     } else if (previousSibling.type === opts.typeItem) {
         return previousSibling;
-    } else {
-        return null;
     }
+    return null;
 }
 
-module.exports = getPreviousItem;
+export default getPreviousItem;
