@@ -1,4 +1,4 @@
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const selectedBlock = value.document.getDescendant('_selection_key');
     change.collapseToStartOf(selectedBlock);
@@ -14,4 +14,4 @@ module.exports = function(plugin, change) {
     );
 
     return change;
-};
+}
