@@ -1,6 +1,6 @@
-const expect = require('expect');
+import expect from 'expect';
 
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const selectedBlock = value.document.getDescendant('_selection_key');
     change.collapseToStartOf(selectedBlock);
@@ -20,4 +20,4 @@ module.exports = function(plugin, change) {
     expect(change.value.selection.anchorOffset).toEqual(0);
     expect(change.value.selection.isCollapsed).toBe(true);
     return change;
-};
+}
