@@ -1,5 +1,4 @@
-
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const selectedBlock = value.document.getDescendant('_selection_key');
     change.collapseToStartOf(selectedBlock);
@@ -13,4 +12,4 @@ module.exports = function(plugin, change) {
         change,
         {}
     );
-};
+}
