@@ -1,6 +1,3 @@
-const Slate = require('slate');
-
-module.exports = function(plugin, change) {
-    const schema = Slate.Schema.create({plugins: [plugin]});
-    return change.setValue({ schema }).normalize();
-};
+export default function(plugin, change) {
+    return change.normalize();
+}
