@@ -1,8 +1,7 @@
-
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const p = value.document.findDescendant(node => node.type == 'paragraph');
 
     change.collapseToStartOf(p);
     return change.call(plugin.changes.splitListItem);
-};
+}
