@@ -1,11 +1,10 @@
-const expect = require('expect');
+import expect from 'expect';
 
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const selectedBlock = value.document.getDescendant('_selection_key');
 
-    change.collapseToStartOf(selectedBlock)
-          .move(2); // It|em 1
+    change.collapseToStartOf(selectedBlock).move(2); // It|em 1
 
     plugin.changes.splitListItem(change);
 
@@ -24,4 +23,4 @@ module.exports = function(plugin, change) {
     });
 
     return change;
-};
+}
