@@ -1,21 +1,22 @@
-const expect = require('expect');
+import expect from 'expect';
 
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const selectedBlock = value.document.getDescendant('_selection_key');
 
-    const initial = change.value.change({ save: false }).moveToRangeOf(selectedBlock);
+    const initial = change.value
+        .change({ save: false })
+        .moveToRangeOf(selectedBlock);
     const initialText = initial.value.startBlock.text;
     const initialSelection = initial.value.selection;
 
     const newChange = initial.value.change();
 
-    newChange.call(plugin.changes.increaseItemDepth)
-          .undo();
+    newChange.call(plugin.changes.increaseItemDepth).undo();
 
     // Back to previous cursor position
     expect(newChange.value.startBlock.text).toEqual(initialText);
     expect(newChange.value.selection.toJS()).toEqual(initialSelection.toJS());
 
     return newChange;
-};
+}
