@@ -1,9 +1,11 @@
-const expect = require('expect');
+import expect from 'expect';
 
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const selectedBlock = value.document.getDescendant('_selection_key');
-    const initial = change.value.change({ save: false }).moveToRangeOf(selectedBlock).value;
+    const initial = change.value
+        .change({ save: false })
+        .moveToRangeOf(selectedBlock).value;
     const initialText = initial.startBlock.text;
     const initialSelection = initial.selection;
     const toTest = initial.change();
@@ -14,4 +16,4 @@ module.exports = function(plugin, change) {
     expect(toTest.value.selection.toJS()).toEqual(initialSelection.toJS());
 
     return toTest;
-};
+}
