@@ -1,8 +1,7 @@
-
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const { value } = change;
     const selectedBlock = value.document.getDescendant('_selection_key');
     change.moveToRangeOf(selectedBlock);
 
     return change.call(plugin.changes.unwrapList);
-};
+}
