@@ -1,5 +1,4 @@
-
-module.exports = function(plugin, change) {
+export default function(plugin, change) {
     const data = { style: { listStyleType: 'decimal' } };
     return change.call(plugin.changes.wrapInList, 'ol_list', data);
-};
+}
